[Decomp][Object] Cohesion split: extract update building to ObjectUpdate.cpp
