+++ src/game/Object/Object.cpp
@@ -202,679 +202,17 @@ void Object::SetObjectScale(float newScale)
     SetFloatValue(OBJECT_FIELD_SCALE_X, newScale);
 }
 
-/**
- * @brief Force immediate update transmission to all viewers
- *
- * Sends all pending update changes immediately rather than waiting
- * for the next update tick. This is used for urgent updates that
- * must be visible immediately (e.g., combat state changes).
- *
- * The method builds update data for all nearby players and sends
- * it immediately, then removes the object from the pending update list.
- */
-void Object::SendForcedObjectUpdate()
-{
-    if (!m_inWorld || !m_objectUpdated)
-    {
-        return;
-    }
 
-    UpdateDataMapType update_players;
 
-    BuildUpdateData(update_players);
-    RemoveFromClientUpdateList();
 
-    WorldPacket packet;                                     // here we allocate a std::vector with a size of 0x10000
-    for (UpdateDataMapType::iterator iter = update_players.begin(); iter != update_players.end(); ++iter)
-    {
-        iter->second.BuildPacket(&packet);
-        iter->first->GetSession()->SendPacket(&packet);
-        packet.clear();                                     // clean the string
-    }
-}
 
-/**
- * @brief Build create update block for player
- * @param data Update data buffer
- * @param target Target player
- *
- * Builds the update packet data needed to create this object
- * for the specified player. Includes movement data and
- * all update field values.
- */
-void Object::BuildCreateUpdateBlockForPlayer(UpdateData* data, Player* target) const
-{
-    if (!target)
-    {
-        return;
-    }
 
-    uint8  updatetype   = UPDATETYPE_CREATE_OBJECT;
-    uint8 updateFlags  = m_updateFlag;
 
-    /** lower flag1 **/
-    if (target == this)                                     // building packet for yourself
-    {
-        updateFlags |= UPDATEFLAG_SELF;
-    }
 
-    if (updateFlags & UPDATEFLAG_HAS_POSITION)
-    {
-        // UPDATETYPE_CREATE_OBJECT2 dynamic objects, corpses...
-        if (isType(TYPEMASK_DYNAMICOBJECT) || isType(TYPEMASK_CORPSE) || isType(TYPEMASK_PLAYER))
-        {
-            updatetype = UPDATETYPE_CREATE_OBJECT2;
-        }
 
-        // UPDATETYPE_CREATE_OBJECT2 for pets...
-        if (target->GetPetGuid() == GetObjectGuid())
-        {
-            updatetype = UPDATETYPE_CREATE_OBJECT2;
-        }
 
-        // UPDATETYPE_CREATE_OBJECT2 for some gameobject types...
-        if (isType(TYPEMASK_GAMEOBJECT))
-        {
-            switch (((GameObject*)this)->GetGoType())
-            {
-                case GAMEOBJECT_TYPE_TRAP:
-                case GAMEOBJECT_TYPE_DUEL_ARBITER:
-                case GAMEOBJECT_TYPE_FLAGSTAND:
-                case GAMEOBJECT_TYPE_FLAGDROP:
-                    updatetype = UPDATETYPE_CREATE_OBJECT2;
-                    break;
-                case GAMEOBJECT_TYPE_TRANSPORT:
-                    updateFlags |= UPDATEFLAG_TRANSPORT;
-                    break;
-                default:
-                    break;
-            }
-        }
 
-        if (isType(TYPEMASK_UNIT))
-        {
-            if (((Unit*)this)->getVictim())
-            {
-                updateFlags |= UPDATEFLAG_HAS_ATTACKING_TARGET;
-            }
-        }
-    }
 
-    // DEBUG_LOG("BuildCreateUpdate: update-type: %u, object-type: %u got updateFlags: %X", updatetype, m_objectTypeId, updateFlags);
-
-    ByteBuffer& buf = data->GetBuffer();
-    buf << uint8(updatetype);
-    buf << GetPackGUID();
-    buf << uint8(m_objectTypeId);
-
-    BuildMovementUpdate(&buf, updateFlags);
-
-    UpdateMask updateMask;
-    updateMask.SetCount(m_valuesCount);
-    _SetCreateBits(&updateMask, target);
-    BuildValuesUpdate(updatetype, &buf, &updateMask, target);
-    data->AddUpdateBlock();
-}
-
-/**
- * @brief Send create update to player
- * @param player Target player
- *
- * Sends the create update packet to the specified player,
- * causing the object to appear in their game world.
- */
-void Object::SendCreateUpdateToPlayer(Player* player)
-{
-    // send create update to player
-    UpdateData upd;
-    WorldPacket packet;
-
-    BuildCreateUpdateBlockForPlayer(&upd, player);
-    upd.BuildPacket(&packet);
-    player->GetSession()->SendPacket(&packet);
-}
-
-/**
- * @brief Build values update block for player
- * @param data Update data buffer
- * @param target Target player
- *
- * Builds the update packet data for changed field values
- * to send to the specified player.
- */
-void Object::BuildValuesUpdateBlockForPlayer(UpdateData* data, Player* target) const
-{
-    ByteBuffer& buf = data->GetBuffer();
-
-    buf << uint8(UPDATETYPE_VALUES);
-    buf << GetPackGUID();
-
-    UpdateMask updateMask;
-    updateMask.SetCount(m_valuesCount);
-
-    _SetUpdateBits(&updateMask, target);
-    BuildValuesUpdate(UPDATETYPE_VALUES, &buf, &updateMask, target);
-
-    data->AddUpdateBlock();
-}
-
-/**
- * @brief Build out of range update block
- * @param data Update data buffer
- *
- * Adds this object's GUID to the out-of-range list,
- * indicating it should be removed from the client's view.
- */
-void Object::BuildOutOfRangeUpdateBlock(UpdateData* data) const
-{
-    data->AddOutOfRangeGUID(GetObjectGuid());
-}
-
-/**
- * @brief Destroy object for player
- * @param target Target player
- *
- * Sends a destroy packet to the specified player,
- * removing this object from their game world.
- */
-void Object::DestroyForPlayer(Player* target) const
-{
-    MANGOS_ASSERT(target);
-
-    WorldPacket data(SMSG_DESTROY_OBJECT, 8);
-    data << GetObjectGuid();
-    target->GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Build movement update block
- * @param data Byte buffer to write to
- * @param updateFlags Update flags
- *
- * Builds the movement data portion of the update packet.
- * Includes position, orientation, movement flags, and speeds
- * for living objects, or just position for static objects.
- */
-void Object::BuildMovementUpdate(ByteBuffer* data, uint8 updateFlags) const
-{
-    *data << uint8(updateFlags);                            // update flags
-
-    // 0x20
-    if (updateFlags & UPDATEFLAG_LIVING)
-    {
-        Unit* unit = ((Unit*)this);
-
-        if (GetTypeId() == TYPEID_PLAYER)
-        {
-            Player* player = ((Player*)unit);
-            if (player->GetTransport())
-            {
-                player->m_movementInfo.AddMovementFlag(MOVEFLAG_ONTRANSPORT);
-            }
-            else
-            {
-                player->m_movementInfo.RemoveMovementFlag(MOVEFLAG_ONTRANSPORT);
-            }
-        }
-
-        // Update movement info time
-        unit->m_movementInfo.UpdateTime(GameTime::GetGameTimeMS());
-        // Write movement info
-        unit->m_movementInfo.Write(*data);
-
-        // Unit speeds
-        *data << float(unit->GetSpeed(MOVE_WALK));
-        *data << float(unit->GetSpeed(MOVE_RUN));
-        *data << float(unit->GetSpeed(MOVE_RUN_BACK));
-        *data << float(unit->GetSpeed(MOVE_SWIM));
-        *data << float(unit->GetSpeed(MOVE_SWIM_BACK));
-        *data << float(unit->GetSpeed(MOVE_FLIGHT));
-        *data << float(unit->GetSpeed(MOVE_FLIGHT_BACK));
-        *data << float(unit->GetSpeed(MOVE_TURN_RATE));
-
-        // 0x08000000
-        if (unit->m_movementInfo.GetMovementFlags() & MOVEFLAG_SPLINE_ENABLED)
-        {
-            Movement::PacketBuilder::WriteCreate(*unit->movespline, *data);
-        }
-    }
-    // 0x40
-    else if (updateFlags & UPDATEFLAG_HAS_POSITION)
-    {
-        // 0x02
-        if (updateFlags & UPDATEFLAG_TRANSPORT && ((GameObject*)this)->GetGoType() == GAMEOBJECT_TYPE_MO_TRANSPORT)
-        {
-            *data << float(0);
-            *data << float(0);
-            *data << float(0);
-            *data << float(((WorldObject*)this)->GetOrientation());
-        }
-        else
-        {
-            *data << float(((WorldObject*)this)->GetPositionX());
-            *data << float(((WorldObject*)this)->GetPositionY());
-            *data << float(((WorldObject*)this)->GetPositionZ());
-            *data << float(((WorldObject*)this)->GetOrientation());
-        }
-    }
-
-    // 0x8
-    if (updateFlags & UPDATEFLAG_LOWGUID)
-    {
-        switch (GetTypeId())
-        {
-            case TYPEID_OBJECT:
-            case TYPEID_ITEM:
-            case TYPEID_CONTAINER:
-            case TYPEID_GAMEOBJECT:
-            case TYPEID_DYNAMICOBJECT:
-            case TYPEID_CORPSE:
-                *data << uint32(GetGUIDLow());              // GetGUIDLow()
-                break;
-            case TYPEID_UNIT:
-                *data << uint32(0x0000000B);                // unk, can be 0xB or 0xC
-                break;
-            case TYPEID_PLAYER:
-                if (updateFlags & UPDATEFLAG_SELF)
-                {
-                    *data << uint32(0x00000015);            // unk, can be 0x15 or 0x22
-                }
-                else
-                {
-                    *data << uint32(0x00000008);            // unk, can be 0x7 or 0x8
-                }
-                break;
-            default:
-                *data << uint32(0x00000000);                // unk
-                break;
-        }
-    }
-
-    // 0x10
-    if (updateFlags & UPDATEFLAG_HIGHGUID)
-    {
-        switch (GetTypeId())
-        {
-            case TYPEID_OBJECT:
-            case TYPEID_ITEM:
-            case TYPEID_CONTAINER:
-            case TYPEID_GAMEOBJECT:
-            case TYPEID_DYNAMICOBJECT:
-            case TYPEID_CORPSE:
-                *data << uint32(GetObjectGuid().GetHigh()); // GetGUIDHigh()
-                break;
-            default:
-                *data << uint32(0x00000000);                // unk
-                break;
-        }
-    }
-
-    // 0x4
-    if (updateFlags & UPDATEFLAG_HAS_ATTACKING_TARGET)      // packed guid (current target guid)
-    {
-        if (((Unit*)this)->getVictim())
-        {
-            *data << ((Unit*)this)->getVictim()->GetPackGUID();
-        }
-        else
-        {
-            data->appendPackGUID(0);
-        }
-    }
-
-    // 0x2
-    if (updateFlags & UPDATEFLAG_TRANSPORT)
-    {
-        *data << uint32(GameTime::GetGameTimeMS());           // ms time
-    }
-}
-
-/**
- * @brief Build values update data
- * @param updatetype Update type (create or values)
- * @param data Byte buffer to write to
- * @param updateMask Update mask indicating which fields changed
- * @param target Target player
- *
- * Builds the actual field value data for the update packet.
- * Handles special cases for gameobjects and units.
- */
-void Object::BuildValuesUpdate(uint8 updatetype, ByteBuffer* data, UpdateMask* updateMask, Player* target) const
-{
-    if (!target)
-    {
-        return;
-    }
-
-    bool IsActivateToQuest = false;
-    bool IsPerCasterAuraState = false;
-
-    if (updatetype == UPDATETYPE_CREATE_OBJECT || updatetype == UPDATETYPE_CREATE_OBJECT2)
-    {
-        if (isType(TYPEMASK_GAMEOBJECT) && !((GameObject*)this)->IsTransport())
-        {
-            if (((GameObject*)this)->ActivateToQuest(target) || target->isGameMaster())
-            {
-                IsActivateToQuest = true;
-            }
-
-            updateMask->SetBit(GAMEOBJECT_DYN_FLAGS);
-        }
-        else if (isType(TYPEMASK_UNIT))
-        {
-            if (((Unit*)this)->HasAuraState(AURA_STATE_CONFLAGRATE))
-            {
-                IsPerCasterAuraState = true;
-                updateMask->SetBit(UNIT_FIELD_AURASTATE);
-            }
-        }
-    }
-    else                                                    // case UPDATETYPE_VALUES
-    {
-        if (isType(TYPEMASK_GAMEOBJECT) && !((GameObject*)this)->IsTransport())
-        {
-            if (((GameObject*)this)->ActivateToQuest(target) || target->isGameMaster())
-            {
-                IsActivateToQuest = true;
-            }
-
-            updateMask->SetBit(GAMEOBJECT_DYN_FLAGS);
-            updateMask->SetBit(GAMEOBJECT_ANIMPROGRESS);
-        }
-        else if (isType(TYPEMASK_UNIT))
-        {
-            if (((Unit*)this)->HasAuraState(AURA_STATE_CONFLAGRATE))
-            {
-                IsPerCasterAuraState = true;
-                updateMask->SetBit(UNIT_FIELD_AURASTATE);
-            }
-        }
-    }
-
-    MANGOS_ASSERT(updateMask && updateMask->GetCount() == m_valuesCount);
-
-    *data << (uint8)updateMask->GetBlockCount();
-    data->append(updateMask->GetMask(), updateMask->GetLength());
-
-    // 2 specialized loops for speed optimization in non-unit case
-    if (isType(TYPEMASK_UNIT))                              // unit (creature/player) case
-    {
-        for (uint16 index = 0; index < m_valuesCount; ++index)
-        {
-            if (updateMask->GetBit(index))
-            {
-                if (index == UNIT_NPC_FLAGS)
-                {
-                    uint32 appendValue = m_uint32Values[index];
-
-                    if (GetTypeId() == TYPEID_UNIT)
-                    {
-                        if (!target->canSeeSpellClickOn((Creature*)this))
-                        {
-                            appendValue &= ~UNIT_NPC_FLAG_SPELLCLICK;
-                        }
-
-                        if (appendValue & UNIT_NPC_FLAG_TRAINER)
-                        {
-                            if (!((Creature*)this)->IsTrainerOf(target, false))
-                            {
-                                appendValue &= ~(UNIT_NPC_FLAG_TRAINER | UNIT_NPC_FLAG_TRAINER_CLASS | UNIT_NPC_FLAG_TRAINER_PROFESSION);
-                            }
-                        }
-
-                        if (appendValue & UNIT_NPC_FLAG_STABLEMASTER)
-                        {
-                            if (target->getClass() != CLASS_HUNTER)
-                            {
-                                appendValue &= ~UNIT_NPC_FLAG_STABLEMASTER;
-                            }
-                        }
-                    }
-
-                    *data << uint32(appendValue);
-                }
-                else if (index == UNIT_FIELD_AURASTATE)
-                {
-                    if (IsPerCasterAuraState)
-                    {
-                        // IsPerCasterAuraState set if related pet caster aura state set already
-                        if (((Unit*)this)->HasAuraStateForCaster(AURA_STATE_CONFLAGRATE, target->GetObjectGuid()))
-                        {
-                            *data << m_uint32Values[index];
-                        }
-                        else
-                        {
-                            *data << (m_uint32Values[index] & ~(1 << (AURA_STATE_CONFLAGRATE - 1)));
-                        }
-                    }
-                    else
-                    {
-                        *data << m_uint32Values[index];
-                    }
-                }
-                // FIXME: Some values at server stored in float format but must be sent to client in uint32 format
-                else if (index >= UNIT_FIELD_BASEATTACKTIME && index <= UNIT_FIELD_RANGEDATTACKTIME)
-                {
-                    // convert from float to uint32 and send
-                    *data << uint32(m_floatValues[index] < 0 ? 0 : m_floatValues[index]);
-                }
-
-                // there are some float values which may be negative or can't get negative due to other checks
-                else if ((index >= PLAYER_FIELD_NEGSTAT0    && index <= PLAYER_FIELD_NEGSTAT4) ||
-                         (index >= PLAYER_FIELD_RESISTANCEBUFFMODSPOSITIVE  && index <= (PLAYER_FIELD_RESISTANCEBUFFMODSPOSITIVE + 6)) ||
-                         (index >= PLAYER_FIELD_RESISTANCEBUFFMODSNEGATIVE  && index <= (PLAYER_FIELD_RESISTANCEBUFFMODSNEGATIVE + 6)) ||
-                         (index >= PLAYER_FIELD_POSSTAT0    && index <= PLAYER_FIELD_POSSTAT4))
-                {
-                    *data << uint32(m_floatValues[index]);
-                }
-
-                // Gamemasters should be always able to select units - remove not selectable flag
-                else if (index == UNIT_FIELD_FLAGS && target->isGameMaster())
-                {
-                    *data << (m_uint32Values[index] & ~UNIT_FLAG_NOT_SELECTABLE);
-                }
-                /* Hide loot animation for players that aren't permitted to loot the corpse */
-                else if (index == UNIT_DYNAMIC_FLAGS && GetTypeId() == TYPEID_UNIT)
-                {
-                    uint32 send_value = m_uint32Values[index];
-
-                    /* Initiate pointer to creature so we can check loot */
-                    if (Creature* my_creature = (Creature*)this)
-                    {
-                        /* If the creature is NOT fully looted */
-                        if (!my_creature->loot.isLooted())
-                        {
-                            /* If the lootable flag is NOT set */
-                            if (!(send_value & UNIT_DYNFLAG_LOOTABLE))
-                            {
-                                /* Update it on the creature */
-                                my_creature->SetFlag(UNIT_DYNAMIC_FLAGS, UNIT_DYNFLAG_LOOTABLE);
-                                /* Update it in the packet */
-                                send_value = send_value | UNIT_DYNFLAG_LOOTABLE;
-                            }
-                        }
-                    }
-                    /* If we're not allowed to loot the target, destroy the lootable flag */
-                    if (!target->isAllowedToLoot((Creature*)this))
-                    {
-                        if (send_value & UNIT_DYNFLAG_LOOTABLE)
-                        {
-                            send_value = send_value & ~UNIT_DYNFLAG_LOOTABLE;
-                        }
-                    }
-
-                    /* If we are allowed to loot it and mob is tapped by us, destroy the tapped flag */
-                    bool is_tapped = target->IsTappedByMeOrMyGroup((Creature*)this);
-
-                    /* If the creature has tapped flag but is tapped by us, remove the flag */
-                    if (send_value & UNIT_DYNFLAG_TAPPED && is_tapped)
-                    {
-                        send_value = send_value & ~UNIT_DYNFLAG_TAPPED;
-                    }
-
-                    *data << send_value;
-                }
-                else                                        // Unhandled index, just send
-                {
-                    // send in current format (float as float, uint32 as uint32)
-                    *data << m_uint32Values[index];
-                }
-            }
-        }
-    }
-    else if (isType(TYPEMASK_GAMEOBJECT))                   // gameobject case
-    {
-        for (uint16 index = 0; index < m_valuesCount; ++index)
-        {
-            if (updateMask->GetBit(index))
-            {
-                // send in current format (float as float, uint32 as uint32)
-                if (index == GAMEOBJECT_DYN_FLAGS)
-                {
-                    // GAMEOBJECT_TYPE_DUNGEON_DIFFICULTY can have lo flag = 2
-                    //      most likely related to "can enter map" and then should be 0 if can not enter
-
-                    if (IsActivateToQuest)
-                    {
-                        switch (((GameObject*)this)->GetGoType())
-                        {
-                            case GAMEOBJECT_TYPE_QUESTGIVER:
-                                *data << uint16(GO_DYNFLAG_LO_ACTIVATE);
-                                *data << uint16(0);
-                                break;
-                            case GAMEOBJECT_TYPE_CHEST:
-                            case GAMEOBJECT_TYPE_GENERIC:
-                            case GAMEOBJECT_TYPE_SPELL_FOCUS:
-                            case GAMEOBJECT_TYPE_GOOBER:
-                                *data << uint16(GO_DYNFLAG_LO_ACTIVATE | GO_DYNFLAG_LO_SPARKLE);
-                                *data << uint16(0);
-                                break;
-                            default:
-                                *data << uint32(0);         // unknown, not happen.
-                                break;
-                        }
-                    }
-                    else
-                    {
-                        // disable quest object
-                        *data << uint32(0);
-                    }
-                }
-                else
-                {
-                    *data << m_uint32Values[index];          // other cases
-                }
-            }
-        }
-    }
-    else                                                    // other objects case (no special index checks)
-    {
-        for (uint16 index = 0; index < m_valuesCount; ++index)
-        {
-            if (updateMask->GetBit(index))
-            {
-                // send in current format (float as float, uint32 as uint32)
-                *data << m_uint32Values[index];
-            }
-        }
-    }
-}
-
-/**
- * @brief Clear update mask
- * @param remove If true, remove from client update list
- *
- * Clears all changed value flags and optionally removes
- * the object from the pending update list.
- */
-void Object::ClearUpdateMask(bool remove)
-{
-    if (m_uint32Values)
-    {
-        for (uint16 index = 0; index < m_valuesCount; ++index)
-        {
-            m_changedValues[index] = false;
-        }
-    }
-
-    if (m_objectUpdated)
-    {
-        if (remove)
-        {
-            RemoveFromClientUpdateList();
-        }
-        m_objectUpdated = false;
-    }
-}
-
-/**
- * @brief Load values from data string
- * @param data Data string to load from
- * @return True if successful
- *
- * Loads update field values from a character data string.
- * Used when loading objects from database.
- */
-bool Object::LoadValues(const char* data)
-{
-    if (!m_uint32Values)
-    {
-        _InitValues();
-    }
-
-    Tokens tokens = StrSplit(data, " ");
-
-    if (tokens.size() != m_valuesCount)
-    {
-        return false;
-    }
-
-    Tokens::iterator iter;
-    int index;
-    for (iter = tokens.begin(), index = 0; index < m_valuesCount; ++iter, ++index)
-    {
-        m_uint32Values[index] = atol((*iter).c_str());
-    }
-
-    return true;
-}
-
-/**
- * @brief Set update bits in mask
- * @param updateMask Update mask to modify
- * @param target Target player (unused)
- *
- * Sets bits in the update mask for all fields that have changed.
- */
-void Object::_SetUpdateBits(UpdateMask* updateMask, Player* /*target*/) const
-{
-    for (uint16 index = 0; index < m_valuesCount; ++index)
-    {
-        if (m_changedValues[index])
-        {
-            updateMask->SetBit(index);
-        }
-    }
-}
-
-/**
- * @brief Set create bits in mask
- * @param updateMask Update mask to modify
- * @param target Target player (unused)
- *
- * Sets bits in the update mask for all non-zero fields.
- * Used when creating a new object for a player.
- */
-void Object::_SetCreateBits(UpdateMask* updateMask, Player* /*target*/) const
-{
-    for (uint16 index = 0; index < m_valuesCount; ++index)
-    {
-        if (GetUInt32Value(index) != 0)
-        {
-            updateMask->SetBit(index);
-        }
-    }
-}
 
 /**
  * @brief Set signed 32-bit value
